Move header stuff to sidebar
