+++ package.json
@@ -23,10 +23,10 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.2",
     "@testing-library/react": "^10.4.7",
-    "@testing-library/user-event": "^12.0.17",
-    "@types/jest": "^26.0.7",
+    "@testing-library/user-event": "^12.1.0",
+    "@types/jest": "^26.0.8",
     "@types/node": "^14.0.27",
-    "@types/react": "^16.9.43",
+    "@types/react": "^16.9.44",
     "@types/react-dom": "^16.9.8",
     "@types/react-redux": "^7.1.9",
     "cross-env": "^7.0.2",
