+++ src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom"
 import { Provider } from "react-redux"
 import React, { Fragment } from "react"
-import { Global, css } from "@emotion/core"
+import { Global } from "@emotion/core"
 // import * as serviceWorker from "./serviceWorker"
 import { fal } from "@fortawesome/pro-light-svg-icons"
 import { fas } from "@fortawesome/pro-solid-svg-icons"
@@ -9,6 +9,7 @@ import { fab } from "@fortawesome/free-brands-svg-icons"
 import { fad } from "@fortawesome/pro-duotone-svg-icons"
 import { far } from "@fortawesome/pro-regular-svg-icons"
 import { library } from "@fortawesome/fontawesome-svg-core"
+import { globalStyles } from "./styled"
 import store from "./store"
 import App from "./App"
 
@@ -18,29 +19,7 @@ library.add(fab, fad, fal, far, fas)
 ReactDOM.render(
   // <React.StrictMode>
   <Fragment>
-    <Global
-      styles={css`
-        @import url("https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;400&family=Tomorrow:wght@400;500&display=swap");
-        * {
-          box-sizing: border-box;
-        }
-        html,
-        body {
-          padding: 0;
-          margin: 0;
-          width: 100vw;
-          height: 100vh;
-          font-family: Roboto, sans-serif;
-          font-style: normal;
-          font-weight: 400;
-          font-size: 16px;
-          background-color: #000;
-        }
-        button {
-          font-family: Tomorrow, sans-serif;
-        }
-      `}
-    />
+    <Global styles={globalStyles} />
     <Provider store={store}>
       <App />
     </Provider>
